Refactor area.js
